fix: 修复 Issue2/4/6 - SQLite查询、事务边界、Double转BigDecimal

+++ src/main/java/com/aiconnecting/dto/CouponGenerateRequest.java
@@ -3,6 +3,7 @@ package com.aiconnecting.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 public class CouponGenerateRequest {
 
     @NotNull(message = "积分不能为空")
-    private Double credits;
+    private BigDecimal credits;
 
     private Integer maxUses;
