+++ src/main/java/com/aiconnecting/dto/TokenRequest.java
@@ -1,13 +1,14 @@
 package com.aiconnecting.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class TokenRequest {
     private String name;
     private Long quota;
-    private Double credits;
+    private BigDecimal credits;
     private LocalDateTime expiredAt;
     private String allowedModels;
     private Integer rateLimit;
